pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whoopsync"
version = "0.1.0"
description = "Get your Whoop data locally and do stuff with it"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "python-dotenv",
    "pydantic",
    "sqlalchemy",
    "pytest"
]

[tool.black]
line-length = 88

[tool.mypy]
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
